[COM-2] Made separate file for endpoint protocol
